@@ -10,40 +10,32 @@
   },
   "dependencies": {
     "@babel/helper-regex": "^7.10.5",
-    "axios": "^0.21.1",
-    "bootstrap": "^4.6.0",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.6.1",
+    "axios": "^0.24.0",
+    "bootstrap": "4.6.1",
+    "react": "^17.0.2",
+    "react-bootstrap": "1.6.1",
     "react-confirm-alert": "^2.7.0",
-    "react-dom": "^17.0.1",
-    "react-google-recaptcha-v3": "^1.7.1",
-    "react-icons": "^4.1.0",
-    "react-redux": "^7.2.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.2",
-    "redux": "^4.0.5",
-    "redux-promise": "^0.6.0",
-    "redux-thunk": "^2.3.0",
+    "react-dom": "^17.0.2",
+    "react-google-recaptcha-v3": "^1.9.7",
+    "react-icons": "^4.3.1",
+    "react-router": "^6.0.2",
+    "react-router-dom": "^6.0.2",
+    "react-scripts": "^4.0.3",
     "socket.io-client": "^4.3.2",
-    "styled-components": "^5.2.1"
+    "styled-components": "^5.3.3"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.20",
-    "@types/node": "^14.14.25",
-    "@types/react": "^17.0.1",
-    "@types/react-bootstrap": "^0.32.25",
-    "@types/react-dom": "^17.0.0",
+    "@types/jest": "^27.0.2",
+    "@types/node": "^16.11.7",
+    "@types/react": "^17.0.35",
+    "@types/react-bootstrap": "^0.32.28",
+    "@types/react-dom": "^17.0.11",
     "@types/react-icons": "^3.0.0",
-    "@types/react-redux": "^7.1.16",
-    "@types/react-router": "^5.1.11",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/redux": "^3.6.31",
-    "@types/redux-promise": "^0.5.29",
-    "@types/redux-thunk": "^2.1.32",
+    "@types/react-router": "^5.1.17",
+    "@types/react-router-dom": "^5.3.2",
     "@types/socket.io-client": "^3.0.0",
-    "@types/styled-components": "^5.1.7",
-    "typescript": "^4.1.3"
+    "@types/styled-components": "^5.1.15",
+    "typescript": "^4.4.4"
   },
   "eslintConfig": {
     "extends": "react-app"
