@@ -1,53 +1,48 @@
 import { Col, Row, Dropdown } from 'react-bootstrap'
 import { isMobile, timeConverter } from '../../services/functions'
 
-export const Col3 = (props:any) => {
+export const Col3 = (props: any) => {
 
     const { vivienda, cambiarEstado } = props
 
     return (
     
-    <Col xs={12} md={3} style={{margin: isMobile ? 'auto' : '0 30px'}}>
+    <Col xs={12} md={3} style={{ margin: isMobile ? 'auto' : '0 30px' }}>
 
-        <Row style={{textAlign:"center", height:"30%", margin: isMobile ? '50 auto' : '20px auto 0 auto'}}>
+        <Row style={{ textAlign: 'center', height: '30%', margin: isMobile ? '50 auto' : '20px auto 0 auto' }}>
 
-            <Dropdown style={{width:'100%', margin: isMobile ? '25px auto' : '30px auto'}}>
+            <Dropdown style={{ width: '100%', margin: isMobile ? '25px auto' : '30px auto' }}>
 
                 <Dropdown.Toggle id="dropdown-basic" variant={vivienda.variante}
-                    style={{width:'80%', border:'1px solid black'}}
+                    style={{ width: '80%', border: '1px solid black' }}
                 >
                     {vivienda.estado}
                 </Dropdown.Toggle>
 
                 <Dropdown.Menu>
                     
-                    <Dropdown.Item onClick={
-                        () => {cambiarEstado(vivienda.inner_id, "No predicado", vivienda.noAbonado, vivienda.asignado)}
-                    }>
+                    <Dropdown.Item
+                        onClick={() => cambiarEstado(vivienda.inner_id, "No predicado", vivienda.noAbonado, vivienda.asignado)}>
                         No predicado
                     </Dropdown.Item>
                     
-                    <Dropdown.Item onClick={
-                        () => {cambiarEstado(vivienda.inner_id, "Contestó", vivienda.noAbonado, vivienda.asignado)}
-                    }>
+                    <Dropdown.Item
+                        onClick={() => cambiarEstado(vivienda.inner_id, "Contestó", vivienda.noAbonado, vivienda.asignado)}>
                         Contestó
                     </Dropdown.Item>
                     
-                    <Dropdown.Item onClick={
-                        () => {cambiarEstado(vivienda.inner_id, "No contestó", vivienda.noAbonado, vivienda.asignado)}
-                    }>
+                    <Dropdown.Item
+                        onClick={() => cambiarEstado(vivienda.inner_id, "No contestó", vivienda.noAbonado, vivienda.asignado)}>
                         No contestó
                     </Dropdown.Item>
                     
-                    <Dropdown.Item onClick={
-                        () => {cambiarEstado(vivienda.inner_id, "A dejar carta", vivienda.noAbonado, vivienda.asignado)}
-                    }>
+                    <Dropdown.Item
+                        onClick={() => cambiarEstado(vivienda.inner_id, "A dejar carta", vivienda.noAbonado, vivienda.asignado)}>
                         A dejar carta
                     </Dropdown.Item>
                     
-                    <Dropdown.Item onClick={
-                        () => {cambiarEstado(vivienda.inner_id, "No llamar", vivienda.noAbonado, vivienda.asignado)}
-                    }>
+                    <Dropdown.Item
+                        onClick={() => cambiarEstado(vivienda.inner_id, "No llamar", vivienda.noAbonado, vivienda.asignado)}>
                         No llamar
                     </Dropdown.Item>
                 </Dropdown.Menu>
@@ -56,21 +51,19 @@ export const Col3 = (props:any) => {
 
         </Row>
 
-        <Row style={{height:'40%', marginTop:'15px'}}>
+        <Row style={{ height: '40%', marginTop: '15px' }}>
             {vivienda.fechaUlt
             ?
                 <div className="card border-dark mb-3"
                     style={{
-                        maxWidth:'18rem',
-                        backgroundColor:'rgb(214, 214, 214)',
-                        display:
-                            vivienda.estado === "No predicado" // && vivienda.noAbonado===false
-                            ? 'none' : 'block',
-                        margin:'auto'
+                        maxWidth: '18rem',
+                        backgroundColor: 'rgb(214, 214, 214)',
+                        display: vivienda.estado === "No predicado" ? 'none' : 'block',
+                        margin: 'auto'
                     }}
                 >
                     
-                    <div className='card-header' style={{padding:'0.2rem 0.5rem'}}>
+                    <div className='card-header' style={{ padding: '0.2rem 0.5rem' }}>
                         <p className='card-text'>
                             Se llamó el {timeConverter(vivienda.fechaUlt, true)}
                         </p>
