@@ -1,4 +1,4 @@
 export const recaptchaPublicKey = "6LfDIdIZAAAAAElWChHQZq-bZzO9Pu42dt9KANY9"
 
-const local = false
+const local = true
 export let SERVER: string = local ? "http://localhost:4005" : "https://registros-rest-api.herokuapp.com"
