@@ -2,7 +2,7 @@ import { SERVER } from '../config'
 import { getHeaders, getToken } from './functions'
 import { typeUser } from '../models/typesUsuarios'
 
-const base = `${SERVER}/api/users`
+const base: string = `${SERVER}/api/users`
 const setToken = (newToken: string) => localStorage.setItem('token', newToken)
 const clearToken = () => localStorage.removeItem('token')
 
@@ -13,39 +13,39 @@ export const loginService = async (email: string, password: string, recaptchaTok
         body: JSON.stringify({ email, password, recaptchaToken })
     })
     const response: any|null = await request.json()
-    if (response && response.success && response.newtoken) setToken(response.newtoken)
-    return response
-}
-
-export const registerUserService = async (email: string,
-     password: string, group: number, recaptchaToken: string): Promise<object|null> => {
-    const request: any = await fetch(`${base}/register`, {
-        method: 'POST',
-        headers: getHeaders(),
-        body: JSON.stringify({ email, password, group, recaptchaToken })
-    })
-    const response: any = await request.json()
-    if (!response || !response.success) { console.log("Failed /register"); return null }
+    if (response && response.success && response.newToken) setToken(response.newToken)
     return response
 }
 
 export const authUserService = async (): Promise<typeUser|null> => {
     const token: string|null = getToken()
     if (!token) return null
-    const request = await fetch(`${base}/auth`, {
+    const request: any|null = await fetch(`${base}/auth`, {
         method: 'POST',
         headers: getHeaders(),
         body: JSON.stringify({ token })
     })
-    const userData: typeUser|null = await request.json()
-    if (!userData || !userData.isAuth) { console.log("Failed /auth"); clearToken(); return null }
-    return userData
+    const response: any|null = await request.json()
+    if (!response || !response.success || !response.user || !response.user.isAuth) clearToken()
+    return response.user
 }
 
 export const autoLoginService = async (): Promise<boolean> => {
     const user: typeUser|null = await authUserService()
     if (!user) return false
     return true
+}
+
+export const registerUserService = async (email: string,
+    password: string, group: number, recaptchaToken: string): Promise<object|null> => {
+   const request: any = await fetch(`${base}/register`, {
+       method: 'POST',
+       headers: getHeaders(),
+       body: JSON.stringify({ email, password, group, recaptchaToken })
+   })
+   const response: any = await request.json()
+   if (!response || !response.success) return null
+   return response
 }
 
 export const getUsersService = async (): Promise<typeUser[]|null> => {
@@ -61,16 +61,17 @@ export const getUsersService = async (): Promise<typeUser[]|null> => {
     return response.users
 }
 
-export const changeDarkModeService = async (newMode: boolean): Promise<object|null> => {
+export const changeDarkModeService = async (newMode: boolean): Promise<boolean> => {
     const token: string|null = getToken()
-    if (!token) return null
+    if (!token) return false
     const request = await fetch(`${base}/change-mode`, {
         method: 'POST',
         headers: getHeaders(),
         body: JSON.stringify({ token, darkMode: newMode })
     })
     const response: any|null = await request.json()
-    return response
+    if (!response || !response.success) return false
+    return true
 }
 
 export const changePswService = async (psw: string, newPsw: string): Promise<object|null> => {
@@ -82,8 +83,9 @@ export const changePswService = async (psw: string, newPsw: string): Promise<obj
         body: JSON.stringify({ token, psw, newPsw })
     })
     const response: any|null = await fetchy.json()
-    if (!response || !response.success) { console.log("Failed /change-psw") }
     if (response && response.success && response.newToken) setToken(response.newToken)
+
+    // ver nuevas respuestas
     return response
 }
 
@@ -96,22 +98,38 @@ export const changePswOtherUserService = async (email: string): Promise<boolean>
         body: JSON.stringify({ token, email })
     })
     const response: any|null = await request.json()
+
+    // ver nuevas respuestas
     if (!response || !response.success) return false
     return true
 }
 
 export const logoutService = async (): Promise<void> => {
+    // const token: string|null = getToken()
+    // if (!token) return;
+    // const request = await fetch(`${base}/logout`, {
+    //     method: 'POST',
+    //     headers: getHeaders(),
+    //     body: JSON.stringify({ token })
+    // })
+    // const response: any|null = await request.json()
+    // if (response && response.success) console.log("Sesión de usuario cerrada con éxito")
+    // else console.log("Algo falló y no cerró sesión bien")
+    clearToken()
+}
+
+export const logoutAllService = async (): Promise<boolean> => {
     const token: string|null = getToken()
-    if (!token) return;
-    const request = await fetch(`${base}/logout`, {
+    if (!token) return false;
+    const request = await fetch(`${base}/logout-all`, {
         method: 'POST',
         headers: getHeaders(),
         body: JSON.stringify({ token })
     })
     const response: any|null = await request.json()
-    if (response && response.success) console.log("Sesión de usuario cerrada con éxito")
-    else console.log("Algo falló y no cerró sesión bien")
-    clearToken()
+    if (!response || !response.success || !response.newToken) return false
+    setToken(response.newToken)
+    return true
 }
 
 export const modifyUserService = async (user_id: string, estado: boolean, role: number, group: number): Promise<typeUser|null> => {
